+++ supabase/migrations/20260225090902_remove_ticket_number.sql
@@ -1,0 +1,3 @@
+alter table "public"."tickets" drop column "ticket_number";
+
+